+++ CHANGELOG.md
@@ -20,6 +20,9 @@
 
 ### Refactored
 
+- [`fc04a6a`](https://github.com/HaydenMCD/jog-journal/commit/fc04a6ada253aa1713adbd245feceb4563b89779) 8 October 2025 Anthony Legg - #8 refactor distance context, use typescript
+
+make it compatible with tanstack router context
 - [`1fb74c8`](https://github.com/HaydenMCD/jog-journal/commit/1fb74c8e1394e82868f700868911fb9ef4a35709) 4 October 2025 Anthony Legg - refactor: add dummy content to home page
 - [`86facbd`](https://github.com/HaydenMCD/jog-journal/commit/86facbd656bab25aa2eefaa0f99b25b3850097cb) 4 October 2025 Anthony Legg - refactor: update `__root.tsx` use distance context
 - [`162215e`](https://github.com/HaydenMCD/jog-journal/commit/162215e489b3288b2483362509db3c69a0513eac) 4 October 2025 Anthony Legg - refactor: delete distance context/provider
@@ -106,6 +109,10 @@ Signed-off-by: Anthony Legg &lt;anthony.legg@outlook.com&gt;
 - [`f67e716`](https://github.com/HaydenMCD/jog-journal/commit/f67e7169c6c1fcf2a88aa1759a2ccebc73376e4d) 9 July 2025 Anthony Legg - #8 revert: remove previous node packages
 
 ## Commits :heavy_check_mark:
+
+- [`fc04a6a`](https://github.com/HaydenMCD/jog-journal/commit/fc04a6ada253aa1713adbd245feceb4563b89779) 8 October 2025 Anthony Legg - #8 refactor distance context, use typescript
+
+make it compatible with tanstack router context
 
 - [`1fb74c8`](https://github.com/HaydenMCD/jog-journal/commit/1fb74c8e1394e82868f700868911fb9ef4a35709) 4 October 2025 Anthony Legg - refactor: add dummy content to home page